+++ packages/styleguide/stories/Labs/Experimental/Atoms/Popover/Popover.examples.tsx
@@ -4,7 +4,6 @@ import { Popover, PopoverProps } from '@codecademy/gamut-labs/src';
 
 export const PopoverExample = (args: PopoverProps) => {
   const [open, setOpen] = useState(false);
-  const [disable, setDisable] = useState(false);
   const activeElRef = useRef<HTMLDivElement>(null);
 
   return (
@@ -12,11 +11,7 @@ export const PopoverExample = (args: PopoverProps) => {
       <div ref={activeElRef}>
         <Button
           onClick={() => {
-            if (disable) {
-              setDisable(false);
-              return;
-            }
-            setOpen(true);
+            setOpen(!open);
           }}
         >
           Open Popover
@@ -28,7 +23,6 @@ export const PopoverExample = (args: PopoverProps) => {
           isOpen={open}
           targetRef={activeElRef}
           onRequestClose={() => setOpen(false)}
-          disableOutsideEvent={() => setDisable(true)}
         >
           <Container>
             <h1>Hooray!</h1>